requirements.txt
Flask==2.3.2
langchain==0.0.237
llama_index==0.7.9
newspaper3k==0.2.8
